self-growth-daily-briefing/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "self-growth-daily-briefing"
version = "0.1.0"
description = "Local-first daily self-growth briefing pipeline with topic ranking, AI writing, and SMTP delivery."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
